azure_function/requirements.txt
ShopifyAPI>=12.5.0
azure-functions
python-dotenv
requests
azure-storage-blob
